# openapi.yaml
openapi: 3.0.3
info:
  title: LucidLayer API
  version: 0.2.0
  description: |
    LucidLayer Offchain API.

    This OpenAPI spec is the source of truth for the actual backend routes.
    It is used to generate SDK clients via Speakeasy.

    Route groups:
    - `/v1/*` — Passports, Match, Run, Receipts, Epochs, Payouts, Compute
    - `/health/*` — System health and dependency checks
    - `/api/agents/*` — AI Agent MMR, Planner, and Orchestrator
servers:
  - url: https://api.lucid.foundation
    description: Production server

# Speakeasy MCP tool generation configuration
...
paths:
...
  /v1/passports:
    post:
      tags: [Passports]
      operationId: lucid_create_passport
      summary: Create a passport
      x-speakeasy-mcp:
        name: lucid_create_passport
        description: Create a new passport (model/compute/tool/dataset/agent).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePassportRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePassportResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    get:
      tags: [Passports]
      operationId: lucid_list_passports
      summary: List passports
      x-speakeasy-mcp:
        name: lucid_list_passports
        description: List passports with filtering, sorting, and pagination.
      parameters:
        - in: query
          name: type
          schema:
            oneOf:
              - $ref: '#/components/schemas/PassportType'
              - type: array
                items: { $ref: '#/components/schemas/PassportType' }
        - in: query
          name: owner
          schema: { type: string }
        - in: query
          name: status
          schema:
            oneOf:
              - $ref: '#/components/schemas/PassportStatus'
              - type: array
                items: { $ref: '#/components/schemas/PassportStatus' }
        - in: query
          name: tags
          description: Comma-separated
          schema: { type: string }
        - in: query
          name: tag_match
          schema: { type: string, enum: [all, any] }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: per_page
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: sort_by
          schema: { type: string, enum: [created_at, updated_at, name] }
        - in: query
          name: sort_order
          schema: { type: string, enum: [asc, desc] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListPassportsResponse'
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/passports/{passport_id}:
    get:
      tags: [Passports]
      operationId: lucid_get_passport
      summary: Get a passport
      x-speakeasy-mcp:
        name: lucid_get_passport
        description: Retrieve a passport by its passport_id.
      parameters:
        - in: path
          name: passport_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPassportResponse'
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    patch:
      tags: [Passports]
      operationId: lucid_update_passport
      summary: Update a passport
      x-speakeasy-mcp:
        name: lucid_update_passport
        description: Update an existing passport by passport_id.
      parameters:
        - in: path
          name: passport_id
          required: true
          schema: { type: string }
        - in: header
          name: X-Owner-Address
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePassportRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPassportResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      tags: [Passports]
      operationId: lucid_delete_passport
      summary: Delete a passport (soft delete)
      x-speakeasy-mcp:
        name: lucid_delete_passport
        description: Soft-delete a passport by passport_id.
      parameters:
        - in: path
          name: passport_id
          required: true
          schema: { type: string }
        - in: header
          name: X-Owner-Address
          required: false
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, deleted]
                properties:
                  success: { type: boolean }
                  deleted: { type: boolean }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/passports/{passport_id}/sync:
    post:
      tags: [Passports]
      operationId: lucid_trigger_passport_sync
      summary: Trigger on-chain sync for a passport
      x-speakeasy-mcp:
        name: lucid_trigger_passport_sync
        description: Trigger on-chain sync for a passport.
      parameters:
        - in: path
          name: passport_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success: { type: boolean }
                  on_chain_pda: { type: string, nullable: true }
                  on_chain_tx: { type: string, nullable: true }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/passports/pending-sync:
    get:
      tags: [Passports]
      operationId: lucid_list_passports_pending_sync
      summary: Get passports pending sync
      x-speakeasy-mcp:
        name: lucid_list_passports_pending_sync
        description: List passports that are pending on-chain sync.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, count, passports]
                properties:
                  success: { type: boolean }
                  count: { type: integer }
                  passports:
                    type: array
                    items: { $ref: '#/components/schemas/Passport' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/passports/stats:
    get:
      tags: [Passports]
      operationId: lucid_get_passport_stats
      summary: Passport statistics
      x-speakeasy-mcp:
        name: lucid_get_passport_stats
        description: Get passport statistics.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PassportStatsResponse'
        '500': { $ref: '#/components/responses/InternalServerError' }

  # ===========================================================================
  # Share Tokens
  # ===========================================================================
...
  /v1/models:
    get:
      tags: [Passports]
      operationId: lucid_search_models
      summary: Search model passports
      x-speakeasy-mcp:
        name: lucid_search_models
        description: Search model passports (ModelMeta filters).
      parameters:
        - in: query
          name: runtime
          schema: { type: string }
        - in: query
          name: format
          schema: { type: string }
        - in: query
          name: max_vram
          schema: { type: integer }
        - in: query
          name: available
          description: "Tri-state filter: 'true' returns only models that can serve inference (healthy compute or API-hosted), 'false' returns only unavailable models (missing compute), omit for all models"
          schema: { type: string, enum: ['true', 'false'] }
        - in: query
          name: owner
          schema: { type: string }
        - in: query
          name: tags
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: per_page
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, models, pagination]
                properties:
                  success: { type: boolean }
                  models:
                    type: array
                    items: { $ref: '#/components/schemas/Passport' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/compute:
    get:
      tags: [Compute]
      operationId: lucid_search_compute
      summary: Search compute passports
      x-speakeasy-mcp:
        name: lucid_search_compute
        description: Search compute passports (ComputeMeta filters).
      parameters:
        - in: query
          name: regions
          schema: { type: string }
        - in: query
          name: runtimes
          schema: { type: string }
        - in: query
          name: provider_type
          schema: { type: string }
        - in: query
          name: min_vram
          schema: { type: integer }
        - in: query
          name: gpu
          schema: { type: string }
        - in: query
          name: owner
          schema: { type: string }
        - in: query
          name: tags
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: per_page
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, compute, pagination]
                properties:
                  success: { type: boolean }
                  compute:
                    type: array
                    items: { $ref: '#/components/schemas/Passport' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /v1/tools:
    get:
      tags: [Passports]
      operationId: lucid_list_tools
      summary: List tool passports
      x-speakeasy-mcp:
        name: lucid_list_tools
        description: List active tool passports with optional filtering.
      parameters:
        - in: query
          name: owner
          schema: { type: string }
        - in: query
          name: tags
          description: Comma-separated
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: per_page
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [success, tools, pagination]
                properties:
                  success: { type: boolean }
                  tools:
                    type: array
                    items: { $ref: '#/components/schemas/Passport' }
                  pagination: { $ref: '#/components/schemas/Pagination' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnprocessableEntity:
      description: Unprocessable Entity
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    PassportType:
      type: string
      enum: [model, compute, tool, dataset, agent, voice, other]
    PassportStatus:
      type: string
      enum: [active, deprecated, revoked]
...
    CreatePassportRequest:
      type: object
      required: [type, owner, metadata]
      properties:
        type: { $ref: '#/components/schemas/PassportType' }
        owner: { type: string }
        metadata:
          type: object
          additionalProperties: true
        name: { type: string }
        description: { type: string }
        version: { type: string }
        tags:
          type: array
          items: { type: string }
...
    CreatePassportResponse:
      type: object
      required: [success, passport_id, passport]
      properties:
        success: { type: boolean }
        passport_id: { type: string }
        passport: { $ref: '#/components/schemas/Passport' }
...
    ListPassportsResponse:
      type: object
      required: [success, passports, pagination]
      properties:
        success: { type: boolean }
        passports:
          type: array
          items: { $ref: '#/components/schemas/Passport' }
        pagination: { $ref: '#/components/schemas/Pagination' }
...